+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ---
 
+#### 16.12.2018
+
+* [YouTubeData] Reduced the amount of API requests
+
 #### 15.12.2018
 
 * [Favourites] Removing favourite on hold and removed bottom bar from PlayerActivity